+++ requirements.txt
@@ -1,36 +1,10 @@
-branca==0.4.2
-Brotli==1.0.9
-certifi==2020.12.5
-chardet==4.0.0
-click==7.1.2
-cycler==0.10.0
+dash_html_components==1.1.2
 dash==1.19.0
-dash-bootstrap-components==0.11.3
-dash-core-components==1.15.0
-dash-html-components==1.1.2
-dash-renderer==1.9.0
-dash-table==4.11.2
-Flask==1.1.2
-Flask-Compress==1.9.0
-folium==0.12.1
-future==0.18.2
-idna==2.10
-itsdangerous==1.1.0
-Jinja2==2.11.3
-kiwisolver==1.3.1
-MarkupSafe==1.1.1
-matplotlib==3.3.4
-numpy==1.20.1
-pandas==1.2.3
-Pillow==8.1.2
 plotly==4.14.3
-pyaml==20.4.0
-pyparsing==2.4.7
-python-dateutil==2.8.1
-pytz==2021.1
+dash_bootstrap_components==0.12.0
+dash_core_components==1.15.0
+pandas==1.2.3
+requests==2.23.0
+haversine==2.3.0
+python_dateutil==2.8.1
 PyYAML==5.4.1
-requests==2.25.1
-retrying==1.3.3
-six==1.15.0
-urllib3==1.26.3
-Werkzeug==1.0.1
