+++ style.py
@@ -1,0 +1,42 @@
+# File that contains CSS constants for Dashboard Application.
+main_colors = {
+    'main-blue': '#343a40',
+    'medium-blue-grey': 'rgb(77, 79, 91)',
+    'superdark-green': 'rgb(41, 56, 55)',
+    'dark-green': 'rgb(57, 81, 85)',
+    'medium-green': 'rgb(93, 113, 120)',
+    'light-green': 'rgb(186, 218, 212)',
+    'pink-red': 'rgb(255, 101, 131)',
+    'dark-pink-red': 'rgb(247, 80, 99)',
+    'white': 'rgb(251, 251, 252)',
+    'light-grey': 'rgb(208, 206, 206)'
+}
+
+main_container = {
+    'margin-top': '30px',
+}
+
+s_card_box = {
+    "maxHeight": "150px",
+    "overflow": "auto",
+    "font-size":" 9px",
+}
+
+s_button = {
+    'margin-left': '15px',
+    'min-width': '100px',
+    'float': 'left',
+    'font-weight': ' 400',
+    'color':  main_colors['white'],
+    'text-align': ' center',
+    'vertical-align': ' middle',
+    '-webkit-user-select': ' none',
+    '-moz-user-select': ' none',
+    '-ms-user-select': ' none',
+    'user-select': ' none',
+    'border': ' 1px solid transparent',
+    'font-size': ' 1rem',
+    'border-radius': ' .25rem',
+    'background-color': main_colors['main-blue'],
+    'border-color': '#343a40',
+}
